feat: heatmap=> densityHeatmap  matrix=> heatmap
