+++ src/plots/density-heatmap/components/background.ts
