+++ src/plots/density-heatmap/components/index.ts
