+++ src/plots/density-heatmap/components/legend.ts
