+++ src/plots/density-heatmap/event.ts
@@ -2,8 +2,9 @@ import { assign } from '@antv/util';
 import { getEventMap, EVENT_MAP, IEventmap, onEvent } from '../../util/event';
 
 const componentMap = {
-  Rect: 'polygon',
+  Heatmap: 'heatmap',
   LegendLabel: 'legend-label',
+  HeatmapBackground: 'heatmap-background',
 };
 
 const SHAPE_EVENT_MAP: IEventmap = getEventMap(componentMap);
