+++ src/plots/heatmap/component/label.ts
