+++ src/plots/heatmap/component/legend.ts
