+++ src/plots/heatmap/layer.ts
@@ -1,118 +1,129 @@
-import { deepMix, each, has } from '@antv/util';
+import { deepMix, valuesOfKey, each } from '@antv/util';
+import BBox from '../../util/bbox';
+import { getScale } from '@antv/scale';
 import { registerPlotType } from '../../base/global';
 import { LayerConfig } from '../../base/layer';
 import ViewLayer, { ViewConfig } from '../../base/view-layer';
-import { getComponent } from '../../components/factory';
-import { getGeom } from '../../geoms/factory';
-import { extractScale } from '../../util/scale';
-import '../../geoms/heatmap/linear';
-import { HeatmapLegendConfig } from './components/legend';
-import { HeatmapBackgroundConfig } from './components/background';
-import { getPlotComponents } from './components';
-import * as EventParser from './event';
-
-interface PointStyle {
-  lineDash?: number[];
-  lineWidth?: number;
-  opacity?: string;
-  fillStyle?: string;
-  strokeStyle?: string;
-}
+import { MatrixLegendConfig } from './component/legend';
+import { getRectPath, getCirclePath, getCircleCurve } from './shape';
+import { getPlotComponents } from './component';
 
 export interface HeatmapViewConfig extends ViewConfig {
-  colorField: string;
-  radius?: number;
-  intensity?: number;
-  point?: {
-    visible?: boolean;
-    shape?: string;
-    size?: number;
-    color?: string;
-    style?: PointStyle;
-  };
-  legend?: HeatmapLegendConfig;
-  background?: HeatmapBackgroundConfig;
+  sizeField?: string;
+  colorField?: string;
+  shapeSize?: number[];
+  shapeType?: string;
+  color?: string[];
+  legend?: MatrixLegendConfig;
 }
 
 export interface HeatmapLayerConfig extends HeatmapViewConfig, LayerConfig {}
 
 export default class HeatmapLayer<T extends HeatmapLayerConfig = HeatmapLayerConfig> extends ViewLayer<T> {
-  public type: string = 'heatmap';
-  protected plotComponents: any[] = [];
-
   public static getDefaultOptions(): any {
     return deepMix({}, super.getDefaultOptions(), {
+      shapeType: 'rect',
+      legend: {
+        visible: true,
+        position: 'right-center',
+      },
+      tooltip: {
+        shared: false,
+        showCrosshairs: false,
+        showMarkers: false,
+      },
       xAxis: {
         visible: true,
-        autoHideLabel: true,
-        autoRotateLabel: true,
-        autoRotateTitle: false,
+        gridAlign: 'center',
         grid: {
-          visible: false,
-        },
-        line: {
           visible: true,
         },
         tickLine: {
           visible: true,
         },
-        label: {
-          visible: true,
+        line: {
+          visible: false,
         },
-        title: {
-          visible: true,
-          offset: 12,
-        },
+        autoRotateLabel: true,
       },
       yAxis: {
         visible: true,
-        autoHideLabel: true,
-        autoRotateLabel: false,
-        autoRotateTitle: true,
+        gridAlign: 'center',
         grid: {
-          visible: false,
-        },
-        line: {
           visible: true,
+          align: 'center',
         },
         tickLine: {
           visible: true,
         },
-        label: {
-          visible: true,
-        },
-        title: {
-          visible: true,
-          offset: 12,
-        },
+        autoRotateLabel: true,
       },
-      tooltip: {
+      color: ['#9ae3d5', '#66cdbb', '#e7a744', '#f1e066', '#f27664', '#e7c1a2'],
+      label: {
         visible: true,
-        crosshairs: {
-          type: 'cross',
-          style: {
-            lineWidth: 2,
-          },
+        adjustColor: true,
+        adjustPosition: true,
+        offset: 0,
+        style: {
+          stroke: 'rgba(255,255,255,0)',
+          lineWidth: 0,
         },
       },
-      legend: {
-        visible: true,
-        position: 'bottom',
-      },
-      color: [
-        'rgba(33,102,172,0)',
-        'rgb(103,169,207)',
-        'rgb(209,229,240)',
-        'rgb(253,219,199)',
-        'rgb(239,138,98)',
-        'rgb(178,24,43)',
-      ],
     });
   }
+
+  public type: string = 'matrix';
+  protected gridSize: number[] = [];
+  protected plotComponents: any[] = [];
 
   public afterRender() {
     this.renderPlotComponents();
     super.afterRender();
+  }
+
+  public changeShape(type: string) {
+    if (this.options.shapeType === type) {
+      return;
+    }
+    this.options.shapeType = type;
+    if (type === 'rect') {
+      const shapes = this.getShapes();
+      this.circleToRect(shapes);
+    } else if (type === 'circle') {
+      const shapes = this.getShapes();
+      this.rectToCircle(shapes);
+    }
+  }
+
+  public mappingSize(field: string) {
+    if (this.options.sizeField && this.options.sizeField === field) {
+      return;
+    }
+    this.options.sizeField = field;
+    // 创建scale
+    const values = valuesOfKey(this.options.data, field);
+    const min = Math.min(...values);
+    const max = Math.max(...values);
+    const LinearScale = getScale('linear');
+    const scale = new LinearScale({
+      min,
+      max,
+    });
+    const shapes = this.getShapes();
+    if (this.options.shapeType === 'rect') {
+      this.rectSizeMapping(shapes, scale, field);
+    } else if (this.options.shapeType === 'circle') {
+      this.circleSizeMapping(shapes, scale, field);
+    }
+  }
+
+  public disableMappingSize() {
+    const shapes = this.getShapes();
+    if (this.options.shapeType === 'rect') {
+      this.rectDisableSizeMapping(shapes);
+    } else if (this.options.shapeType === 'circle') {
+      this.circleDisableSizeMapping(shapes);
+    }
   }
 
   public destroy() {
@@ -122,32 +133,36 @@ export default class HeatmapLayer<T extends HeatmapLayerConfig = HeatmapLayerCon
     super.destroy();
   }
 
-  protected scale() {
-    const props = this.options;
-    const scales = {};
-    /** 配置x-scale */
-    scales[props.xField] = {};
-    if (has(props, 'xAxis')) {
-      extractScale(scales[props.xField], props.xAxis);
-    }
-    /** 配置y-scale */
-    scales[props.yField] = {};
-    if (has(props, 'yAxis')) {
-      extractScale(scales[props.yField], props.yAxis);
-    }
-    this.setConfig('scales', scales);
-    super.scale();
+  protected geometryParser() {
+    return '';
   }
 
   protected coord() {}
 
-  protected geometryParser(dim, type) {
-    return 'heatmap';
+  protected legend() {
+    this.setConfig('legends', false);
   }
 
   protected addGeometry() {
-    const config = {
-      type: 'linearheatmap',
+    this.gridSize = this.getGridSize();
+    if (this.options.shapeType === 'rect') {
+      const rect = this.addRect();
+      this.setConfig('geometry', rect);
+    } else {
+      const circle = this.addCircle();
+      this.setConfig('geometry', circle);
+    }
+  }
+
+  protected addRect() {
+    // 如果用户设置了size，将size数值转换为[0,1]区间
+    const size = [0.3, 0.9];
+    if (this.options.shapeSize) {
+      size[0] = this.options.shapeSize[0] / this.gridSize[0];
+      size[1] = this.options.shapeSize[1] / this.gridSize[1];
+    }
+    const rect: any = {
+      type: 'polygon',
       position: {
         fields: [this.options.xField, this.options.yField],
       },
@@ -155,71 +170,219 @@ export default class HeatmapLayer<T extends HeatmapLayerConfig = HeatmapLayerCon
         fields: [this.options.colorField],
         values: this.options.color,
       },
-      cfg: {
-        intensity: this.options.intensity,
-        radius: this.options.radius,
+      shape: {
+        values: ['rect'],
       },
-    } as any;
-
-    if (this.options.radius) {
-      config.radius = this.options.radius;
-    }
-
-    if (this.options.intensity) {
-      config.intensity = this.options.intensity;
-    }
-
-    this.setConfig('geometry', config);
-
-    this.addPoint();
-  }
-
-  protected addPoint() {
-    const props = this.options;
-    const defaultConfig = { visible: false, size: 0 };
-    if (props.point && props.point.visible) {
-      props.point = deepMix(defaultConfig, props.point);
+      label: false,
+    };
+    if (this.options.sizeField) {
+      rect.size = {
+        fields: [this.options.sizeField],
+        values: size,
+      };
     } else {
-      props.point = defaultConfig;
+      rect.size = {
+        values: [1],
+      };
     }
-    const point = getGeom('point', 'guide', {
-      plot: this,
-    });
-    point.active = false;
-    // point.label = this.extractLabel();
-    this.setConfig('geometry', point);
+    return rect;
   }
 
-  protected extractLabel() {
-    const props = this.options;
-    const label = props.label;
-    if (label && label.visible === false) {
-      return false;
+  protected addCircle() {
+    let size = [0.3, 0.9];
+    if (this.options.shapeSize) {
+      size = this.options.shapeSize;
+    } else {
+      size[0] = this.gridSize[0] * size[0] * 0.5;
+      size[1] = this.gridSize[1] * size[1] * 0.5;
     }
-    const labelConfig = getComponent('label', {
-      plot: this,
-      labelType: 'scatterLabel',
-      fields: [props.xField, props.yField],
-      position: 'middle',
-      offset: 0,
-      ...label,
+    const circle: any = {
+      type: 'point',
+      position: {
+        fields: [this.options.xField, this.options.yField],
+      },
+      color: {
+        fields: [this.options.colorField],
+        values: this.options.color,
+      },
+      shape: {
+        values: ['curvePoint'],
+      },
+      label: false,
+    };
+    if (this.options.sizeField) {
+      circle.size = {
+        fields: [this.options.sizeField],
+        values: size,
+      };
+    } else {
+      circle.size = {
+        values: [Math.min(this.gridSize[0], this.gridSize[1]) * 0.5 * 0.9],
+      };
+    }
+    return circle;
+  }
+
+  private getGridSize() {
+    if (this.options.padding === 'auto') {
+      return [0, 0];
+    } else {
+      const viewRange = this.getViewRange();
+      const { padding, xField, yField, data } = this.options;
+      const width = viewRange.width - padding[1] - padding[3];
+      const height = viewRange.height - padding[0] - padding[2];
+      const xCount = valuesOfKey(data, xField).length;
+      const yCount = valuesOfKey(data, yField).length;
+      return [width / xCount, height / yCount];
+    }
+  }
+
+  private circleToRect(shapes) {
+    const gridSize = this.gridSize;
+    each(shapes, (shape) => {
+      const { x, y, size } = shape.get('origin');
+      let sizeRatio = (size * 2) / Math.min(gridSize[0], gridSize[1]);
+      if (!this.options.sizeField) {
+        sizeRatio = 1;
+      }
+      const curvePath = getCircleCurve(x, y, size);
+      const rectPath = getRectPath(x, y, gridSize[0], gridSize[1], sizeRatio);
+      shape.stopAnimate();
+      shape.attr('path', curvePath);
+      shape.animate(
+        {
+          path: rectPath,
+        },
+        500,
+        'easeLinear'
+      );
     });
-    return labelConfig;
   }
 
-  protected legend() {
-    this.setConfig('legends', false);
+  private rectToCircle(shapes) {
+    each(shapes, (shape) => {
+      const coord = shape.get('coord');
+      const { points } = shape.get('origin');
+      const ps = [];
+      each(points, (p) => {
+        ps.push(coord.convertPoint(p));
+      });
+      const bbox = shape.getBBox();
+      const width = bbox.width;
+      const height = bbox.height;
+      const centerX = bbox.minX + width / 2;
+      const centerY = bbox.minY + height / 2;
+      const offsetRatio = this.options.sizeField ? 1 : 0.9;
+      const curvePath = getCircleCurve(centerX, centerY, (Math.min(width, height) / 2) * offsetRatio);
+      const circlePath = getCirclePath(centerX, centerY, (Math.min(width, height) / 2) * offsetRatio);
+      shape.stopAnimate();
+      shape.animate(
+        {
+          path: curvePath,
+        },
+        500,
+        'easeLinear',
+        () => {
+          shape.attr('path', circlePath);
+        }
+      );
+    });
   }
 
-  protected parseEvents(eventParser) {
-    super.parseEvents(EventParser);
+  private rectSizeMapping(shapes, scale, field) {
+    each(shapes, (shape) => {
+      const data = shape.get('origin').data;
+      const ratio = 0.3 + scale.scale(data[field]) * 0.6;
+      shape.get('origin').size = ratio;
+      const bbox = shape.getBBox();
+      const width = bbox.width;
+      const height = bbox.height;
+      const centerX = bbox.minX + width / 2;
+      const centerY = bbox.minY + height / 2;
+      const path = getRectPath(centerX, centerY, width, height, ratio);
+      shape.stopAnimate();
+      shape.animate(
+        {
+          path: path,
+        },
+        500,
+        'easeLinear'
+      );
+    });
+  }
+
+  private circleSizeMapping(shapes, scale, field) {
+    each(shapes, (shape) => {
+      const data = shape.get('origin').data;
+      const ratio = 0.3 + scale.scale(data[field]) * 0.6;
+      const { x, y, size } = shape.get('origin');
+      const path = getCirclePath(x, y, size * ratio);
+      shape.get('origin').size = size * ratio;
+      shape.stopAnimate();
+      shape.animate(
+        {
+          path: path,
+        },
+        500,
+        'easeLinear'
+      );
+    });
+  }
+
+  private circleDisableSizeMapping(shapes) {
+    this.options.sizeField = null;
+    each(shapes, (shape) => {
+      const { x, y } = shape.get('origin');
+      const size = Math.min(this.gridSize[0], this.gridSize[1]) * 0.9;
+      shape.get('origin').size = size / 2;
+      const path = getCirclePath(x, y, size / 2);
+      shape.stopAnimate();
+      shape.animate(
+        {
+          path: path,
+        },
+        500,
+        'easeLinear'
+      );
+    });
+  }
+
+  private rectDisableSizeMapping(shapes) {
+    this.options.sizeField = null;
+    each(shapes, (shape) => {
+      const bbox = shape.getBBox();
+      const width = bbox.width;
+      const height = bbox.height;
+      const centerX = bbox.minX + width / 2;
+      const centerY = bbox.minY + height / 2;
+      const path = getRectPath(centerX, centerY, this.gridSize[0], this.gridSize[1], 1);
+      shape.get('origin').size = 1;
+      shape.stopAnimate();
+      shape.animate(
+        {
+          path: path,
+        },
+        500,
+        'easeLinear'
+      );
+    });
+  }
+
+  private getShapes() {
+    const elements = this.view.geometries[0].elements;
+    const shapes = [];
+    each(elements, (ele) => {
+      shapes.push(ele.shape);
+    });
+    return shapes;
   }
 
   protected renderPlotComponents() {
     each(this.plotComponents, (component) => {
       component.destroy();
     });
-    const componentsType = ['legend', 'background'];
+    this.plotComponents = [];
+    const componentsType = ['label', 'legend'];
     each(componentsType, (t) => {
       const cfg = {
         view: this.view,
