+++ src/plots/heatmap/shape.ts
